The final Main program.
Through a lot of time and research, I finally was able to use Cichelli's
method working for the most part. Basically, it just takes the keywords,
creates a hash tables according to those keywords, then runs another
text file through so that it may count the number of each keyword within
the file.