Finished product for my library.
The program takes in a text file, whether it exists or not, and adds
information that can be transcribed as a hashmap. That hashmap can
therein be edited so that more books can be added or removed. Also a
feature of finding certain books is present, as well as a table
displaying the library as a whole. Javadocs are attached, as well as a
lot of notes.